alert manager workable

+++ projects/SmartHouseApp/server/src/main/java/com/balabas/smarthouse/server/entity/alarmV2/IAlarmStateChangeAction.java
@@ -7,8 +7,8 @@ public interface IAlarmStateChangeAction extends IItemAbstract{
 	AlarmState getNewState();
 	AlarmState getOldState();
 	
-	Long getInterval();
-	void setInterval(Long interval);
+	Integer getInterval();
+	void setInterval(Integer interval);
 	
 	boolean accepts(AlarmState oldState, AlarmState newState);
